move integrations registry from Integration class to Instrumentation module
